+++ src/main/java/ru/sumenkov/msf/service/Utility.java
@@ -1,9 +1,6 @@
 package ru.sumenkov.msf.service;
 
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.IOException;
-import java.util.List;
 
 public class Utility {
 
@@ -28,27 +25,27 @@ public class Utility {
         }
     }
 
-    public static boolean allNull(Object[] array) {
-        boolean b = false;
+//    public static boolean allNull(Object[] array) {
+//        boolean b = false;
+//
+//        for (Object i : array) {
+//            if (i != null) {
+//                b = true;
+//                break;
+//            }
+//        }
+//        return b;
+//    }
 
-        for (Object i : array) {
-            if (i != null) {
-                b = true;
-                break;
-            }
-        }
-        return b;
-    }
-
-    public static void closeFiles(List<BufferedReader> inputFiles) {
-        for (BufferedReader br : inputFiles) {
-            try {
-                br.close();
-            } catch (IOException e) {
-                System.out.println("Не смог закрыть поток данных из массива потоков.");
-            }
-        }
-    }
+//    public static void closeFiles(List<BufferedReader> inputFiles) {
+//        for (BufferedReader br : inputFiles) {
+//            try {
+//                br.close();
+//            } catch (IOException e) {
+//                System.out.println("Не смог закрыть поток данных из массива потоков.");
+//            }
+//        }
+//    }
 
     public static void deleteDirectory(File file) {
         if (file.isDirectory()) {
@@ -69,22 +66,22 @@ public class Utility {
         }
     }
 
-    public static void deleteFiles(File file) {
-        if (file.isDirectory()) {
-            File[] contents = file.listFiles();
-            if (contents != null) {
-                for (File f : contents) {
-                    if (f.isFile()) {
-                        if (!f.delete()) {
-                            System.out.printf("Не удалось удалить файл %s\n", file.getName());
-                        }
-                    } else if (f.isDirectory()) {
-                        deleteDirectory(f);
-                    }
-                }
-            } else {
-                System.out.printf("Не удалось найти файл %s\n", file.getName());
-            }
-        }
-    }
+//    public static void deleteFiles(File file) {
+//        if (file.isDirectory()) {
+//            File[] contents = file.listFiles();
+//            if (contents != null) {
+//                for (File f : contents) {
+//                    if (f.isFile()) {
+//                        if (!f.delete()) {
+//                            System.out.printf("Не удалось удалить файл %s\n", file.getName());
+//                        }
+//                    } else if (f.isDirectory()) {
+//                        deleteDirectory(f);
+//                    }
+//                }
+//            } else {
+//                System.out.printf("Не удалось найти файл %s\n", file.getName());
+//            }
+//        }
+//    }
 }
